requirements.txt
Flask==3.1.1
tensorflow==2.19.0
Pillow==11.2.1
numpy==2.1.3
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.2.1

# Google Drive APIs
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
httplib2==0.22.0

# Supabase client
supabase==2.15.3
gotrue==2.12.0
storage3==0.11.3
postgrest==1.0.2

# Other essentials
requests==2.32.4
python-dateutil==2.9.0.post0
protobuf==5.29.5

